Use fair mutex for CallbackCamera
This fixes lock contention that prevented any operations on the camera once capture loop is started by using a fair mutex instead of a regular one.

It's slightly more expensive, but this usecase - trying to access a shared resource that is repeatedly locked/unlocked by someone else - is exactly what it's designed for, and indeed it fixes the deadlock issue.

Fixes #111.