Fix mip_level bug in frame_texture causing crash
The frame_texture method was trying to write to mip_level 1 when the
texture only has 1 mip level (index 0). This causes a crash because of the validation error:
'Unable to select texture mip level 1 out of 1'

Changed mip_level from 1 to 0 to match the texture descriptor which
sets mip_level_count to 1.